+++ CHANGELOG.md
@@ -29,6 +29,12 @@
 * Added `lerp` method to `Hex`
 * Added `rounding_div` method to `Hex` to perform floating point division rounded to the closest coordinate
 * Added `rounding_mul` method to `Hex` to perform floating point multiplication rounded to the closest coordinate
+* Added `cached_rings` and `cached_custom_rings` to `Hex` for rings pre-computation
+
+### Iterator extensions
+
+* Added `average` method for `Hex` iterators to compute their mean value
+* Added `center` method for `Hex` iterators to compute their centroid value
 
 ### Conversions
 
@@ -43,7 +49,10 @@
 * Implemented `RemAssign<Hex>` for `Hex`
 * Implemented `Rem<i32>` for `Hex`
 * Implemented `RemAssign<i32>` for `Hex`
+* Implemented `Sum` for `Hex`
 * `Hex::neighbor` is now `const`
+* `Hex::distance_to` is now `const`
+* `Hex::unsigned_distance_to` is now `const`
 
 ## 0.2.0